Implement ApiClient for all endpoints
Change delete endpoints to only take the id of the object to delete.
Change delete endpoints to use the HTTP DELETE method.

Add protocol objects to the frontend for talking to the backend.
Use kotlin coroutines instead of callbacks for the ApiClient.

Create tests that connect to http://localhost:3000 and tests with the
backend.