Fix small issues
Use ElevatedCard for the device cards.
Fix issue where left and right parts of the card's shadow is cut off.
Clean up the DeviceCard to not use spacers.
Move the delete button into the expanded area in the lower right corner.
Animate the expansion and contraction of the card.

Use a scaffold as the root node of the app.
This provides the bottom bar, background color and other niceties.

Change pagesInfo list to a map.