Fix database is locked error
Write Transactions should be avoided if possible because the lock the
whole database while the transaction is ongoing.

We used it in the create_task endpoint, to check for existance and
ownership of the device the user wants to create a task for.
This has been replaced with a single sql statement that does the same
without a transaction.

This should hopefully fix our flakey tests.