Add background service that runs the scheduling algorithm
This service runs in the background, and runs the scheduling algorithm,
when tasks has changed.
This happens in the following conditions:
- Deleting a device
- Deleting a task
- Creating a task

After one of these things happen a timer is started on the background
service.
This timer is set to 5 minutes.
After 5 minutes it runs the algorithm.
This is done to make sure that it takes at least 5 minutes between runs
of the algorithm, while only running the algorithm when something has
changed.

This also fixes some small off-by-one issues in the NaiveScheduler.